requirements.txt
openai>=1.30.0
anthropic>=0.39.0
matplotlib>=3.8.0
numpy>=1.26.0
scipy>=1.12.0
deap>=1.4.1
python-dotenv>=1.0.0
